hackish rlgl.bi
hackish rlgl.bi implementation. only the declarations works and it need the raylib lib to work well. all the RLGL_STANDALONE and RLGL_IMPLEMENTATION stuff isn translatet (it is a mix of FB and C code to this time so it can be confusing a bit but i know what im doing ;D )